Moved Nouns Fork files to their own directory.
