+++ src/index.ts
@@ -2,9 +2,9 @@ export { EventData, NounsOptions } from "./types";
 export { NounsNymz } from "./contracts/nouns-nymz/NounsNymz";
 export { FederationNounsPool } from "./contracts/federation/FederationNounsPool";
 export { Nouns } from "./contracts/nouns-dao/Nouns";
-export { _NounsForkToken as NounsForkToken } from "./contracts/NounsForkToken";
-export { _NounsForkAuctionHouse as NounsForkAuctionHouse } from "./contracts/NounsForkAuctionHouse";
-export { _NounsFork as NounsFork } from "./contracts/NounsFork";
+export { _NounsForkToken as NounsForkToken } from "./contracts/nouns-fork/NounsForkToken";
+export { _NounsForkAuctionHouse as NounsForkAuctionHouse } from "./contracts/nouns-fork/NounsForkAuctionHouse";
+export { _NounsFork as NounsFork } from "./contracts/nouns-fork/NounsFork";
 export { _Propdates as Propdates } from "./contracts/propdates/Propdates";
 export { LilNouns } from "./contracts/lil-nouns/LilNouns";
 export { IndexerWriter } from "./indexing/IndexerWriter";
